adding check for datatype sizes

+++ pillole/pillole.cpp
@@ -1,11 +1,10 @@
 #include <iostream>
-#include <unordered_map>
 
 using namespace std;
 
-long cache[35][35] = {1};
+unsigned long long cache[35][35] = {1};
 
-long solveN(int n, int m = 0) {
+unsigned long long solveN(int n, int m = 0) {
     return cache[n][m] != 0 ? cache[n][m] :
             cache[n][m] = (n > 0 ? solveN(n - 1, m + 1) : 0) + (m > 0 ? solveN(n, m - 1) : 0);
 }
